Fixed AI, fixed adding players. Working on sync

+++ impact/lib/game/entities/enemy.js
@@ -11,7 +11,7 @@ ig.module(
     collides: ig.Entity.COLLIDES.ACTIVE,
     type: ig.Entity.TYPE.B,
     checkAgainst: ig.Entity.TYPE.A,
-    mob: 'basic',
+    mob: 'EntityEnemy',
     animation: '',
     health: 500,
     pts: 100,
@@ -77,13 +77,17 @@ ig.module(
     	}
     	this.parent();
     },
-    kill: function() {
+    kill: function( removal ) {
+        if(removal !== undefined) {
+          console.log('Despawning');
+          return;
+        };
     	ig.game.increaseScore(100); //adds to score
     	ig.game.addKillCount(); //adds to kill count
     	if (GameInfo.score === 1000) {//Game is WON!
     		ig.game.gameWon();
     	}
-    	socket.emit("score", this.lastHit, this.pts); // Tells the server who killed the guy and the points. 
+    	socket.emit("mob_death", {mob: this.mob, tag: this.tag}); // Tells the server who killed the guy and the points. 
     	this.parent();
     }
   });
